refactor: extract DebugInfoSnapshot.swift from DebugBundleTypes.swift
Splits the `DebugInfoSnapshot` Equatable data value (~49 lines) out of
DebugBundleTypes.swift.

Byte-identical move of lines 3-51. Both files import Foundation.

DebugBundleTypes.swift: 161 → 111 lines. Tests: 667.
